add generic args with support for pe
