+++ libwild/src/diff.rs
@@ -14,7 +14,7 @@ use std::path::PathBuf;
 use std::process::Command;
 
 pub(crate) fn maybe_diff() -> Result {
-    if let Ok(reference_linker) = std::env::var(crate::args::REFERENCE_LINKER_ENV)
+    if let Ok(reference_linker) = std::env::var(crate::args::consts::REFERENCE_LINKER_ENV)
         && let Some(paths) = run_with_linker(&reference_linker)?
     {
         run_diff(&paths)?;
