+++ libwild/src/elf_writer.rs
@@ -4,8 +4,9 @@ use self::elf::NoteProperty;
 use self::elf::get_page_mask;
 use crate::OutputKind;
 use crate::alignment;
-use crate::args::ElfArgs;
-use crate::args::BuildIdOption;
+use crate::args::Args;
+use crate::args::linux::ElfArgs;
+use crate::args::linux::BuildIdOption;
 use crate::bail;
 use crate::debug_assert_bail;
 use crate::elf;
@@ -4228,7 +4229,7 @@ struct DynamicEntryWriter {
 }
 
 struct DynamicEntryInputs<'layout> {
-    args: &'layout ElfArgs,
+    args: &'layout Args<ElfArgs>,
     has_static_tls: bool,
     has_variant_pcs: bool,
     section_layouts: &'layout OutputSectionMap<OutputRecordLayout>,
