+++ libwild/src/input_data.rs
@@ -4,7 +4,8 @@ use crate::archive;
 use crate::archive::ArchiveEntry;
 use crate::archive::ArchiveIterator;
 use crate::archive::EntryMeta;
-use crate::args::ElfArgs;
+use crate::args::Args;
+use crate::args::linux::ElfArgs;
 use crate::args::Input;
 use crate::args::InputSpec;
 use crate::args::Modifiers;
@@ -130,7 +131,7 @@ pub(crate) struct InputLinkerScript<'data> {
 }
 
 struct TemporaryState<'data, O: ObjectFile<'data>> {
-    args: &'data ElfArgs,
+    args: &'data Args<ElfArgs>,
 
     /// Mapping from paths to the index in `files` at which we'll place the result.
     path_to_load_index: Mutex<HashMap<PathBuf, FileLoadIndex>>,
@@ -205,7 +206,7 @@ pub(crate) struct AuxiliaryFiles<'data> {
 }
 
 impl<'data> AuxiliaryFiles<'data> {
-    pub(crate) fn new(args: &'data ElfArgs, inputs_arena: &'data Arena<InputFile>) -> Result<Self> {
+    pub(crate) fn new(args: &'data Args<ElfArgs>, inputs_arena: &'data Arena<InputFile>) -> Result<Self> {
         let resolve_script_path = |path: &Path| -> PathBuf {
             if path.exists() {
                 path.to_owned()
@@ -243,7 +244,7 @@ impl<'data> FileLoader<'data> {
     pub(crate) fn load_inputs<O: ObjectFile<'data>>(
         &mut self,
         inputs: &[Input],
-        args: &'data ElfArgs,
+        args: &'data Args<ElfArgs>,
         plugin: &mut Option<LinkerPlugin<'data>>,
     ) -> Result<LoadedInputs<'data, O>> {
         timing_phase!("Open input files");
@@ -408,7 +409,7 @@ impl<'data> FileLoader<'data> {
 
 fn process_linker_script<'data>(
     input_file: &'data InputFile,
-    args: &ElfArgs,
+    args: &Args<ElfArgs>,
 ) -> Result<LoadedLinkerScript<'data>> {
     let bytes = input_file.data();
     let script = LinkerScript::parse(bytes, &input_file.filename)?;
@@ -698,7 +699,7 @@ fn read_script_data<'data>(
 }
 
 impl Input {
-    fn path(&self, args: &ElfArgs) -> Result<InputPath> {
+    fn path(&self, args: &Args<ElfArgs>) -> Result<InputPath> {
         match &self.spec {
             InputSpec::File(p) => {
                 if self.search_first.is_some()
