+++ libwild/src/output_kind.rs
@@ -1,4 +1,5 @@
-use crate::args::ElfArgs;
+use crate::args::Args;
+use crate::args::linux::ElfArgs;
 use crate::args::RelocationModel;
 use crate::input_data::FileLoader;
 
@@ -10,7 +11,7 @@ pub(crate) enum OutputKind {
 }
 
 impl OutputKind {
-    pub(crate) fn new(args: &ElfArgs, input_data: &FileLoader<'_>) -> OutputKind {
+    pub(crate) fn new(args: &Args<ElfArgs>, input_data: &FileLoader<'_>) -> OutputKind {
         if !args.should_output_executable {
             OutputKind::SharedObject
         } else if args.dynamic_linker.is_some()
