+++ libwild/src/resolution.rs
@@ -4,7 +4,8 @@
 
 use crate::LayoutRules;
 use crate::alignment::Alignment;
-use crate::args::ElfArgs;
+use crate::args::Args;
+use crate::args::linux::ElfArgs;
 use crate::bail;
 use crate::debug_assert_bail;
 use crate::elf::RawSymbolName;
@@ -688,7 +689,7 @@ pub(crate) struct ResolvedLtoInput {
 fn assign_section_ids<'data, O: ObjectFile<'data>>(
     resolved: &mut [ResolvedGroup<'data, O>],
     output_sections: &mut OutputSections<'data>,
-    args: &ElfArgs,
+    args: &Args<ElfArgs>,
 ) {
     timing_phase!("Assign section IDs");
 
@@ -1123,7 +1124,7 @@ impl<'data, O: ObjectFile<'data>> ResolvedDynamic<'data, O> {
 
 fn resolve_sections_for_object<'data, O: ObjectFile<'data>>(
     obj: &mut ResolvedObject<'data, O>,
-    args: &ElfArgs,
+    args: &Args<ElfArgs>,
     allocator: &bumpalo_herd::Member<'data>,
     loaded_metrics: &LoadedMetrics,
     rules: &SectionRules,
@@ -1151,7 +1152,7 @@ fn resolve_section<'data, O: ObjectFile<'data>>(
     input_section_index: SectionIndex,
     input_section: &O::SectionHeader,
     obj: &mut ResolvedObject<'data, O>,
-    args: &ElfArgs,
+    args: &Args<ElfArgs>,
     allocator: &bumpalo_herd::Member<'data>,
     loaded_metrics: &LoadedMetrics,
     rules: &SectionRules,
