+++ libwild/src/subprocess/linux.rs
@@ -1,4 +1,4 @@
-use crate::args::ElfArgs;
+use crate::args::Args;
 use crate::bail;
 use crate::error::Context as _;
 use crate::error::Result;
@@ -20,7 +20,7 @@ use std::ffi::c_void;
 /// # Safety
 /// Must not be called once threads have been spawned. Calling this function from main is generally
 /// the best way to ensure this.
-pub unsafe fn run_in_subprocess(args: ElfArgs) -> ! {
+pub unsafe fn run_in_subprocess(args: Args) -> ! {
     let exit_code = match subprocess_result(args) {
         Ok(code) => code,
         Err(error) => crate::error::report_error_and_exit(&error),
@@ -28,7 +28,7 @@ pub unsafe fn run_in_subprocess(args: ElfArgs) -> ! {
     std::process::exit(exit_code);
 }
 
-fn subprocess_result(args: ElfArgs) -> Result<i32> {
+fn subprocess_result(args: Args) -> Result<i32> {
     let mut fds: [c_int; 2] = [0; 2];
     // create the pipe used to communicate between the parent and child processes - exit on failure
     make_pipe(&mut fds).context("make_pipe")?;
@@ -43,7 +43,7 @@ fn subprocess_result(args: ElfArgs) -> Result<i32> {
             crate::setup_tracing(&args)?;
             let args = args.activate_thread_pool()?;
             let linker = crate::Linker::new();
-            let _outputs = linker.run(&args)?;
+            linker.run(args)?;
             crate::timing::finalise_perfetto_trace()?;
             inform_parent_done(&fds);
             Ok(0)
@@ -51,7 +51,7 @@ fn subprocess_result(args: ElfArgs) -> Result<i32> {
         -1 => {
             // Fork failure in the parent - Fallback to running linker in this process
 
-            crate::run_elf(args)?;
+            crate::run(args)?;
             Ok(0)
         }
         pid => {
