+++ libwild/src/subprocess/mod.rs
@@ -9,8 +9,8 @@ pub use linux::run_in_subprocess;
 pub use windows::run_in_subprocess;
 
 #[cfg(not(any(target_os = "linux", target_os = "windows")))]
-pub unsafe fn run_in_subprocess(args: crate::args::ElfArgs) -> ! {
-    let exit_code = match crate::run_elf(args) {
+pub unsafe fn run_in_subprocess(args: crate::args::Args) -> ! {
+    let exit_code = match crate::run(args) {
         Ok(()) => 0,
         Err(error) => crate::error::report_error_and_exit(&error),
     };
