+++ libwild/src/subprocess_unsupported.rs
@@ -1,7 +1,7 @@
 /// # Safety
 /// See function of the same name in `subprocess.rs`
-pub unsafe fn run_in_subprocess(args: crate::args::ElfArgs) -> ! {
-    let exit_code = match crate::run_elf(args) {
+pub unsafe fn run_in_subprocess(args: crate::args::Args) -> ! {
+    let exit_code = match crate::run(args) {
         Ok(()) => 0,
         Err(error) => {
             eprintln!("{}", error.to_string());
