+++ libwild/src/symbol_db.rs
@@ -4,7 +4,8 @@
 use crate::InputLinkerScript;
 use crate::OutputKind;
 use crate::args;
-use crate::args::ElfArgs;
+use crate::args::Args;
+use crate::args::linux::ElfArgs;
 use crate::bail;
 use crate::elf::RawSymbolName;
 use crate::error;
@@ -71,7 +72,7 @@ use symbolic_demangle::demangle;
 
 #[derive(Debug)]
 pub struct SymbolDb<'data, O: ObjectFile<'data>> {
-    pub(crate) args: &'data ElfArgs,
+    pub(crate) args: &'data Args<ElfArgs>,
 
     pub(crate) groups: Vec<Group<'data, O>>,
 
@@ -321,7 +322,7 @@ impl<'data, O: ObjectFile<'data>> SymbolDb<'data, O> {
     }
 
     pub(crate) fn new(
-        args: &'data ElfArgs,
+        args: &'data Args<ElfArgs>,
         output_kind: OutputKind,
         auxiliary: &AuxiliaryFiles<'data>,
         herd: &'data bumpalo_herd::Herd,
@@ -1393,7 +1394,7 @@ pub(crate) fn is_mapping_symbol_name(name: &[u8]) -> bool {
 fn read_symbols<'data, O: ObjectFile<'data>>(
     version_script: &VersionScript,
     shards: &mut [SymbolWriterShard<'_, '_, 'data, O>],
-    args: &ElfArgs,
+    args: &Args<ElfArgs>,
     export_list: &Option<ExportList<'data>>,
     output_kind: OutputKind,
 ) -> Result<Vec<SymbolLoadOutputs<'data>>> {
@@ -1421,7 +1422,7 @@ fn read_symbols_for_group<'data, O: ObjectFile<'data>>(
     version_script: &VersionScript,
     export_list: &Option<ExportList<'data>>,
     num_buckets: usize,
-    args: &ElfArgs,
+    args: &Args<ElfArgs>,
     output_kind: OutputKind,
 ) -> Result<SymbolLoadOutputs<'data>> {
     verbose_timing_phase!(
@@ -1558,7 +1559,7 @@ fn load_symbols_from_file<'data, O: ObjectFile<'data>>(
     version_script: &VersionScript,
     symbols_out: &mut SymbolWriterShard<'_, '_, 'data, O>,
     outputs: &mut SymbolLoadOutputs<'data>,
-    args: &ElfArgs,
+    args: &Args<ElfArgs>,
     export_list: &Option<ExportList<'data>>,
     output_kind: OutputKind,
 ) -> Result {
@@ -1685,7 +1686,7 @@ trait SymbolLoader<'data, O: ObjectFile<'data>> {
 
 struct RegularObjectSymbolLoader<'a, 'data, O: ObjectFile<'data>> {
     object: &'a O,
-    args: &'a ElfArgs,
+    args: &'a Args<ElfArgs>,
     version_script: &'a VersionScript<'a>,
     archive_semantics: bool,
     lib_name: &'data [u8],
@@ -2070,7 +2071,7 @@ impl<'data> PendingVersionedSymbol<'data> {
 }
 
 /// Decides how many buckets we should use for symbol names.
-fn num_symbol_hash_buckets(args: &ElfArgs) -> usize {
+fn num_symbol_hash_buckets(args: &Args<ElfArgs>) -> usize {
     args.available_threads.get()
 }
